:hammer: Make the builtin editor more self-contained
Rather than spread out the commands and provider for the editor, around the
rest of the app, move all things to do with the editor into its own
self-contained module.
